feat: redesign Vesu market list with card layout
